module/apmhttp: don't test http2 with Go 1.8
http2 no longer builds with Go 1.8, add a build constraint.
